requirements.txt
flask
flask-cors
python-dotenv
psycopg2-binary
google-generativeai